requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database
sqlalchemy>=2.0.28
psycopg2-binary>=2.9.10 
python-dotenv==1.0.1

# AWS SDK
boto3==1.34.59

# PDF Generation
reportlab==4.1.0